+++ glogarch/__init__.py
@@ -1,3 +1,3 @@
 """Glogarch - Graylog Open Archive tool."""
 
-__version__ = "1.10.11"
+__version__ = "1.10.12"